+++ android/app/src/main/kotlin/com/hiddify/hiddify/bg/ServiceNotification.kt
@@ -51,7 +51,7 @@ class ServiceNotification(private val status: MutableLiveData<Status>, private v
         NotificationCompat.Builder(service, notificationChannel)
                 .setShowWhen(false)
                 .setOngoing(true)
-                .setContentTitle("Hiddify Next")
+                .setContentTitle("Hiddify")
                 .setOnlyAlertOnce(true)
                 .setSmallIcon(R.drawable.ic_stat_logo)
                 .setCategory(NotificationCompat.CATEGORY_SERVICE)
@@ -90,7 +90,7 @@ class ServiceNotification(private val status: MutableLiveData<Status>, private v
         }
         service.startForeground(
             notificationId, notificationBuilder
-                .setContentTitle(profileName.takeIf { it.isNotBlank() } ?: "Hiddify Next")
+                .setContentTitle(profileName.takeIf { it.isNotBlank() } ?: "Hiddify")
                 .setContentText(service.getString(contentTextId)).build()
         )
     }