+++ windows/runner/main.cpp
@@ -10,14 +10,14 @@
 int APIENTRY wWinMain(_In_ HINSTANCE instance, _In_opt_ HINSTANCE prev,
                       _In_ wchar_t *command_line, _In_ int show_command) {
   HANDLE hMutexInstance = CreateMutex(NULL, TRUE, L"HiddifyMutex");
-  HWND handle = FindWindowA(NULL, "Hiddify Next");
+  HWND handle = FindWindowA(NULL, "Hiddify");
 
   if (GetLastError() == ERROR_ALREADY_EXISTS) {
     flutter::DartProject project(L"data");
     std::vector<std::string> command_line_arguments = GetCommandLineArguments();
     project.set_dart_entrypoint_arguments(std::move(command_line_arguments));
     FlutterWindow window(project);
-    if (window.SendAppLinkToInstance(L"Hiddify Next")) {
+    if (window.SendAppLinkToInstance(L"Hiddify")) {
       return false;
     }
 
@@ -47,7 +47,7 @@ int APIENTRY wWinMain(_In_ HINSTANCE instance, _In_opt_ HINSTANCE prev,
   FlutterWindow window(project);
   Win32Window::Point origin(10, 10);
   Win32Window::Size size(1280, 720);
-  if (!window.Create(L"Hiddify Next", origin, size)) {
+  if (!window.Create(L"Hiddify", origin, size)) {
     return EXIT_FAILURE;
   }
   window.SetQuitOnClose(true);